requirements.txt
langchain==0.3.18
langchain-community==0.3.17
langchain-ai21==1.0.1
langchain-chroma==0.2.2
pypdf==4.3.1
pandas==2.2.3
datasets==3.3.0
ragas==0.2.13
python-dotenv==1.0.1
streamlit==1.42.0
ragas==0.2.13
datasets==3.3.0
pytest==8.3.4
